Move classes to its own file
