+++ features/mpabout/src/main/java/com/jpp/mpabout/licenses/content/LicenseContent.kt
@@ -1,0 +1,18 @@
+package com.jpp.mpabout.licenses.content
+
+import android.view.View
+/**
+ * Represents the content of the screen.
+ */
+internal data class LicenseContent(
+    val visibility: Int = View.INVISIBLE,
+    val licenseUrl: String = ""
+) {
+    companion object {
+        fun withContent(url: String) =
+            LicenseContent(
+                visibility = View.VISIBLE,
+                licenseUrl = url
+            )
+    }
+}
